Fix sound id variable width bug
Some 64-bit compilers implement C long differently (32-bit vs. 64-bit).  Sound id must be big enough to hold a Python id() int (platform pointer).  Fix sound id C vars to use Uint64 to ensure it is big enough to hold Python int.

+++ mpfmc/core/audio/track_standard.pxd
@@ -52,8 +52,8 @@ ctypedef struct SoundSettings:
     int loops_remaining
     int current_loop
     Uint32 sample_pos
-    long sound_id
-    long sound_instance_id
+    Uint64 sound_id
+    Uint64 sound_instance_id
     int sound_priority
     FadingStatus fading_status
     Uint32 fade_in_steps
@@ -93,8 +93,8 @@ cdef class TrackStandard(Track):
     # destruction.
     cdef TrackStandardState *type_state
 
-    cdef int _get_playing_sound_count(self, int sound_id)
-    cdef list _get_playing_sound_instances(self, int sound_id)
+    cdef int _get_playing_sound_count(self, Uint64 sound_id)
+    cdef list _get_playing_sound_instances(self, Uint64 sound_id)
     cdef int _get_idle_sound_player(self)
     cdef process_notification_message(self, NotificationMessageContainer *notification_message)
     cdef tuple _get_sound_player_with_lowest_priority(self)
